requirements.txt
streamlit==1.25.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
ultralytics==8.0.145
dill==0.3.3
library_name==version

streamlit>=0.84.0
opencv-python>=4.5.3.56
numpy>=1.19.5
Pillow>=8.3.1